Оптимизируем функцию проверки прав пользователей

+++ app/Http/Controllers/Api/V1/AuthorTicketsController.php
@@ -29,17 +29,14 @@ class AuthorTicketsController extends ApiController
      */
     public function store(StoreTicketRequest $request, $author_id)
     {
-        try {
-            // policy
-             $this->isAble('store', Ticket::class);
-
-             return new TicketResource(Ticket::create($request->mappedAttributes([
+        if ($this->isAble('store', Ticket::class)) {
+            return new TicketResource(Ticket::create($request->mappedAttributes([
                 'author' => 'user_id'
-             ])));
-         } catch (AuthorizationException $ex) {
-             return $this->error('You are not authorized to create that resource', 401);
-         }
-    } 
+            ])));
+        }
+
+        return $this->error('You are not authorized to create that resource', 401);
+    }
 
     /**
      * Remove the specified resource from storage.
@@ -51,17 +48,16 @@ class AuthorTicketsController extends ApiController
                             ->where('user_id', $author_id)
                             ->firstOrFail();
 
-            $this->isAble('delete', $ticket);
+            if ($this->isAble('delete', $ticket)) {
+                $ticket->delete();
+                return $this->ok('Ticket successfully deleted');       
+            }
 
-            $ticket->delete();
-            return $this->ok('Ticket successfully deleted'); 
-
+            return $this->error('You are not authorized to delete that resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to delete that resource', 401);
         }
-    } 
+    }
 
     public function replace(ReplaceTicketRequest $request, $author_id,  $ticket_id) {
         // PUT
@@ -70,34 +66,34 @@ class AuthorTicketsController extends ApiController
                             ->where('user_id', $author_id)
                             ->firstOrFail();
 
-            $this->isAble('replace', $ticket);
-
-            $ticket->update($request->mappedAttributes());
-            return new TicketResource($ticket);            
-
+            if ($this->isAble('replace', $ticket)) {
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
+                
+            return $this->error('You are not authorized to update that resource', 401);
+            
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
-    } 
+    }
 
     public function update(UpdateTicketRequest $request, $author_id,  $ticket_id) {
         // PUT
         try {
             $ticket = Ticket::where('id', $ticket_id)
                             ->where('user_id', $author_id)
-                            ->firstOrFail();            
+                            ->firstOrFail();
 
-            $this->isAble('update', $ticket);
+            if ($this->isAble('update', $ticket)) {
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);    
+            }
 
-            $ticket->update($request->mappedAttributes());
-            return new TicketResource($ticket);
-    
+            return $this->error('You are not authorized to update that resource', 401);
+
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
-    }             
+    }            
 }
