+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -28,14 +28,11 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request)
     {
-        try {
-            // policy
-            $this->isAble('store', Ticket::class);
-
+        if ($this->isAble('store', Ticket::class)) {
             return new TicketResource(Ticket::create($request->mappedAttributes()));
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 401);     
         }
+
+        return $this->error('You are not authorized to update that resource', 401);        
     }
 
     /**
@@ -65,17 +62,17 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
-            // policy
-            $this->isAble('update', $ticket);
+            if ($this->isAble('update', $ticket)) {
+                $ticket->update($request->mappedAttributes());
+    
+                return new TicketResource($ticket);
+            }
 
-            $ticket->update($request->mappedAttributes());
+            return $this->error('You are not authorized to update that resource', 401);
 
-            return new TicketResource($ticket);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 401);
-        }        
+        }
     }
 
     /**
@@ -87,11 +84,14 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             // policy
-            $this->isAble('delete', $ticket);
-            
-            $ticket->delete();
+            if ($this->isAble('delete', $ticket)) {
+                $ticket->delete();
 
-            return $this->ok('Ticket successfully deleted');
+                return $this->ok('Ticket successfully deleted');
+            }
+
+            return $this->error('You are not authorized to delete that resource', 401);
+
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot found.', 404);
         }
@@ -103,11 +103,15 @@ class TicketController extends ApiController
             $ticket = Ticket::findOrFail($ticket_id);
 
             // policy
-            $this->isAble('replace', $ticket);
+            if ($this->isAble('replace', $ticket)) {
+                $ticket->update($request->mappedAttributes());
+                return new TicketResource($ticket);
+            }
 
-            $ticket->update($request->mappedAttributes());
+            return $this->error('You are not authorized to update that resource', 401);
 
-            return new TicketResource($ticket);
+
+            
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
         }
